+++ src/Components/UserProfile/Student.js
@@ -12,13 +12,16 @@ import {
 } from "mdb-react-ui-kit";
 import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
-export default function TeacherCom({ teacherData }) {
+
+
+
+export default function TeacherCom({ users }) {
   const [show, setShow] = useState(false);
 
   const handleClose = () => setShow(false);
   const handleShow = () => setShow(true);
   useEffect(() => {
-    console.log(teacherData);
+    console.log(users);
   });
   return (
     <div className="vh-100" style={{ backgroundColor: "#9de2ff" }}>
@@ -37,9 +40,9 @@ export default function TeacherCom({ teacherData }) {
                     /> */}
                   </div>
                   <div className="flex-grow-1 ms-3">
-                    <MDBCardTitle>Name : {teacherData.name}</MDBCardTitle>
-                    <MDBCardTitle>Email : {teacherData.email}</MDBCardTitle>
-                    <MDBCardTitle>Class : {teacherData.class}</MDBCardTitle>
+                    <MDBCardTitle>Name : {users.name}</MDBCardTitle>
+                    <MDBCardTitle>Email : {users.email}</MDBCardTitle>
+                    <MDBCardTitle>Class : {users.class}</MDBCardTitle>
                     <div className="d-flex pt-1">
                       <Button variant="secondary" onClick={handleShow}>
                         View
@@ -57,12 +60,10 @@ export default function TeacherCom({ teacherData }) {
           <Modal.Title>Student Details</Modal.Title>
         </Modal.Header>
         <Modal.Body>
-          <MDBCardTitle>Name : {teacherData.name}</MDBCardTitle>
-          <MDBCardTitle>Email : {teacherData.email}</MDBCardTitle>
-          <MDBCardTitle>Father Name : {teacherData.fatherName}</MDBCardTitle>
-          <MDBCardTitle>Mother Name : {teacherData.MotherName}</MDBCardTitle>
-          <MDBCardTitle>Subject : {teacherData.subject}</MDBCardTitle>
-          <MDBCardTitle>Class : {teacherData.class}</MDBCardTitle>
+          <MDBCardTitle>Name : {users.name}</MDBCardTitle>
+          <MDBCardTitle>Email : {users.email}</MDBCardTitle>
+          {/* <MDBCardTitle>Subject : {teacherData.subject}</MDBCardTitle>
+          <MDBCardTitle>Class : {teacherData.class}</MDBCardTitle> */}
         </Modal.Body>
         <Modal.Footer>
           <Button variant="secondary" onClick={handleClose}>