+++ changelog.md
@@ -1,2 +1,3 @@
  # Added
  - Create User Role , Login by admin , Protected Routes [making-principal-admin]
+ - Routing of UserProfile of student and teacher [userprofile]